abft: add test_abft.py for measuring roundoff error in abft for gemm
This does not use tensorcast (tensorcast does not
implement roundoff error for calculations), but for pytorch
builtin types, it allows collection of abft error due to
fp roundoff. Inputs control the data type used, the size of gemm,
the error tolerance for counting violations, and whether or
not to use modulo arithmetic.