abft: enable excel-friendly cdf graph print for false positive rate
add support for a different method of reporting round-off
errors: fraction of randomized gemm executions that would result
in a false positive ABFT error detection for different threshold
values.

This will output false positive rate data that can be plotted
in a CDF style in excel by selecting a line-based scatter plot,
using the error threshold column for the x axis and the #
false positives column as the y axis. Each data point and/or
gemm size can be added as a different series in the same plot.

set print_cdf=False to disable this and print original format (max
checksum divergence)